Separate Amir's decomposition algorithm as a independent class
1. Create AmirDecomposition class.
2. Simplify Decomposition class.
